+++ packages/airless-core/airless/core/hook/base.py
@@ -1,7 +1,7 @@
 
 import logging
 
-from airless.core.config import get_config
+from airless.core.utils import get_config
 
 
 class BaseHook():