+++ packages/airless-core/airless/core/hook/llm.py
@@ -4,7 +4,7 @@ from airless.core.hook import BaseHook
 
 class LLMHook(BaseHook):
 
-    def __init__(self, model_name, **kwargs):
+    def __init__(self):
         super().__init__()
         self.historic = ''